backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
mysql-connector-python==8.3.0
pytz==2024.1